feat: end-to-end makefile
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch master
# Your branch is up to date with 'origin/master'.
#
# Changes to be committed:
#

modified:   Makefile
#
modified:   pkg/corpus/write.go
#	new
file:   scripts/gather_corpus_dir.sh
#
modified:   scripts/predict/main.go
#
# Untracked files:
#	.vscode/launch.json
#	commit_convention.yml
#	corpus.db
#	db
#	oracle_2_orb.md
#	oracle_3_cookie.md
#	oracle_ascii_1_colums.md
#	oracle_ascii_3_8ball.md
#	pkg/corpus/sql/get_predictions.sql
#	pkg/oracles/README.md
#

+++ pkg/corpus/write.go
@@ -17,7 +17,7 @@ func RegisterOracle(db *sql.DB, oracleName string) (id uint64, err error) {
 	id = DeriveOracleId(oracleName)
 	_, err = db.Exec(
 		"INSERT INTO oracles (id, name) VALUES (?, ?);",
-		id, oracleName)
+		int64(id), oracleName)
 	return id, err
 }
